fixed various errors

+++ Script.sql
@@ -25,10 +25,10 @@ CREATE table proveedores (
 CREATE table productos (
 	codigo bigint primary key,
 	nit_proveedor bigint not null,
-	iva_compra decimal not null,	
+	iva_compra double precision not null,	
 	nombre_producto varchar(40) not null,
-	precio_compra decimal not null,
-	precio_venta decimal not null,
+	precio_compra double precision not null,
+	precio_venta double precision not null,
 	foreign key(nit_proveedor) references proveedores(nit)
 );
 
@@ -36,9 +36,9 @@ CREATE table ventas (
 	codigo bigint primary key,
 	cedula_cliente bigint not null,
 	cedula_usuario bigint not null,
-	iva_venta decimal not null,	
-	total_venta decimal not null,
-	valor_venta decimal not null,
+	iva_venta double precision not null,	
+	total_venta double precision not null,
+	valor_venta double precision not null,
 	foreign key(cedula_cliente) references clientes(cedula),
 	foreign key(cedula_usuario) references usuarios(cedula)
 );
@@ -48,9 +48,9 @@ CREATE table detalle_ventas (
 	codigo_producto bigint not null,
 	codigo_venta bigint not null,
 	cantidad_producto int not null,	
-	valor_total decimal not null,
-	valor_venta decimal not null,
-	valor_iva decimal not null,
+	valor_total double precision not null,
+	valor_venta double precision not null,
+	valor_iva double precision not null,
 	foreign key(codigo_producto) references productos(codigo),
 	foreign key(codigo_venta) references ventas(codigo)
 );
