+++ migrations/20231122215456_productos.sql
@@ -2,9 +2,9 @@
 CREATE table productos (
 	codigo bigint primary key,
 	nit_proveedor bigint not null,
-	iva_compra decimal not null,	
+	iva_compra double precision not null,	
 	nombre_producto varchar(40) not null,
-	precio_compra decimal not null,
-	precio_venta decimal not null,
+	precio_compra double precision not null,
+	precio_venta double precision not null,
 	foreign key(nit_proveedor) references proveedores(nit)
 );
