+++ migrations/20231122215500_ventas.sql
@@ -3,9 +3,10 @@ CREATE table ventas (
 	codigo bigint primary key,
 	cedula_cliente bigint not null,
 	cedula_usuario bigint not null,
-	iva_venta decimal not null,	
-	total_venta decimal not null,
-	valor_venta decimal not null,
+	iva_venta double precision not null,	
+	total_venta double precision not null,
+	valor_venta double precision not null,
 	foreign key(cedula_cliente) references clientes(cedula),
 	foreign key(cedula_usuario) references usuarios(cedula)
 );
+
